requirements.txt
vk_api
python-Levenshtein